Add React navigator and bottom tab drawer
